Refactor: Restructure src/assets/js directory
I've improved the organization and readability of your JavaScript assets:

- I consolidated `index.js` re-exports by removing intermediate index files and updating direct imports in `src/assets/js/index.js`.
- I relocated `analytics.js` to the `utils` directory.
- I renamed the `features` directory to `components` to better reflect its content.
- I updated all relevant import paths in `src/assets/js/index.js` to match the new structure.
