+++ src/assets/js/components/openLinksInPortal.js
