--- src/assets/js/features/index.js
@@ -1,3 +1,0 @@
-export { getHolidays } from './holidays'
-export { menuMobile } from './menu'
-export { openLinksInPortal } from './openLinksInPortal';
