+++ src/assets/js/utils/analytics.js
