--- src/assets/js/utils/index.js
@@ -1,4 +1,0 @@
-export { mutationObserver } from './mutationObserver';
-export { targetBlank } from './targetBlank';
-export { setCurrentYear } from './currentYear';
-export { isMobile } from './isMobile';
